# Fix benchmarking of multiple functions in the same hat file