--- a/hatlib/hat.py
+++ b/hatlib/hat.py
@@ -102,12 +102,9 @@ def hat_description_to_python_function(hat_description: hat_file.HATFile,
 
         launches = func_desc.get("launches")
         if not launches:
-
-            hat_arg_descriptions = func_desc["arguments"]
-            function_name = func_desc["name"]
             hat_library: ctypes.CDLL = hat_details.shared_lib
 
-            def f(function_name, *args):
+            def f(function_name, hat_arg_descriptions, *args):
                 # verify that the (numpy) input args match the description in
                 # the hat file
                 arg_infos = [ArgInfo(d) for d in hat_arg_descriptions]
@@ -122,7 +119,7 @@ def hat_description_to_python_function(hat_description: hat_file.HATFile,
                 # call the function in the hat package
                 hat_library[function_name](*hat_args)
 
-            yield func_name, partial(f, function_name)
+            yield func_name, partial(f, func_desc["name"], func_desc["arguments"])
 
         else:
             device_func = hat_description.get("device_functions",