The capture for function_name is being overwritten if there are multiple functions being benchmarked.

Before the fix:
```
Benchmarking function: matmul_relu_fusion_naive_64a1f4d149d6f346
[Benchmarking] Using 27 input sets, each 3145728 bytes
[Benchmarking] Warming up for 5 iterations...
[Benchmarking] Timing for at least 5s and at least 5 iterations...
[Benchmarking] Mean duration per iteration: 0.02730914s  <-- actually calling matmul_relu_fusion_transformed_5e23ad255420696c

Benchmarking function: matmul_relu_fusion_transformed_5e23ad255420696c
[Benchmarking] Using 27 input sets, each 3145728 bytes
[Benchmarking] Warming up for 5 iterations...
[Benchmarking] Timing for at least 5s and at least 5 iterations...
[Benchmarking] Mean duration per iteration: 0.02665616s
```

After the fix:

```
Benchmarking function: matmul_relu_fusion_naive_64a1f4d149d6f346
[Benchmarking] Using 27 input sets, each 3145728 bytes
[Benchmarking] Warming up for 5 iterations...
[Benchmarking] Timing for at least 5s and at least 5 iterations...
[Benchmarking] Mean duration per iteration: 0.10668768s

Benchmarking function: matmul_relu_fusion_transformed_5e23ad255420696c
[Benchmarking] Using 27 input sets, each 3145728 bytes
[Benchmarking] Warming up for 5 iterations...
[Benchmarking] Timing for at least 5s and at least 5 iterations...
[Benchmarking] Mean duration per iteration: 0.02880869s
```